+++ internal/issuer/surface_test.go
@@ -29,7 +29,7 @@ func TestTheOpenIDSurfaceIsServed(t *testing.T) {
 	if err != nil {
 		t.Fatalf("policy set: %v", err)
 	}
-	iss := issuer.New(issuer.Config{URL: "http://issuer.example", AllowInsecure: true}, set, &fakeDirectory{})
+	iss := issuer.New(issuer.Config{URL: "http://issuer.example", AllowInsecure: true}, set, &fakeDirectory{}, issuer.NewMemoryState())
 
 	storage, err := issuer.NewStorage(iss, fakeVerifier{}, nil, nil, nil)
 	if err != nil {